Add coldata i2c peek and poke to wib_client
Also fixes all warnings in the code, and missing 0 hex offsets for axi modules.

+++ sw/src/femb.h
@@ -36,17 +36,17 @@ public:
     FEMB(int index);
     ~FEMB();
     
-    void fast_cmd(int coldata_idx, uint8_t cmd_code);
+    void fast_cmd(uint8_t coldata_idx, uint8_t cmd_code);
     
-    void i2c_write(int coldata_idx, uint8_t chip_addr, uint8_t reg_page, uint8_t reg_addr, uint8_t data);
-    uint8_t i2c_read(int coldata_idx, uint8_t chip_addr, uint8_t reg_page,  uint8_t reg_addr);
+    void i2c_write(uint8_t coldata_idx, uint8_t chip_addr, uint8_t reg_page, uint8_t reg_addr, uint8_t data);
+    uint8_t i2c_read(uint8_t coldata_idx, uint8_t chip_addr, uint8_t reg_page,  uint8_t reg_addr);
 
 protected:
     int last_coldata_i2c_chip[2];
     io_reg_t coldata_i2c[2];
     io_reg_t coldata_fast_cmd[2];
     
-    void i2c_bugfix(int coldata_idx, uint8_t chip_addr, uint8_t reg_page, uint8_t reg_addr);
+    void i2c_bugfix(uint8_t coldata_idx, uint8_t chip_addr, uint8_t reg_page, uint8_t reg_addr);
     
 };
