+++ sw/src/sensors.cc
@@ -8,7 +8,7 @@
 #include "io_reg.h"
 
 double read_ltc2499_temp(i2c_t *i2c, uint8_t ch) {
-    uint8_t cmd[2] = { 0xB0 | ((ch%2)<<3) | (ch/2), 0x80};
+    uint8_t cmd[2] = { (uint8_t)(0xB0 | ((ch%2)<<3) | (ch/2)), 0x80};
     i2c_write(i2c,0x15,cmd,2);
     usleep(300000);
     uint32_t value;
@@ -35,7 +35,7 @@ void enable_ltc2990(i2c_t *i2c, uint8_t slave) {
 
 //ch 1 to ch 4, Vcc
 uint16_t read_ltc2990_value(i2c_t *i2c, uint8_t slave, uint8_t ch) {
-    uint8_t read_cmd[1] = { 6+(ch-1)*2 };
+    uint8_t read_cmd[1] = { (uint8_t)(6+(ch-1)*2) };
     i2c_write(i2c,slave,read_cmd,1);
     uint16_t value;
     i2c_read(i2c,slave,(uint8_t*)&value,2);
@@ -51,7 +51,7 @@ void enable_ltc2991(i2c_t *i2c, uint8_t slave) {
 
 //ch 1 to ch 4, T_internal, Vcc
 uint16_t read_ltc2991_value(i2c_t *i2c, uint8_t slave, uint8_t ch) {
-    uint8_t read_cmd[1] = { 0xA+(ch-1)*2 };
+    uint8_t read_cmd[1] = { (uint8_t)(0xA+(ch-1)*2) };
     i2c_write(i2c,slave,read_cmd,1);
     uint16_t value;
     i2c_read(i2c,slave,(uint8_t*)&value,2);
